+++ BeeOverlay/PluginController.cs
@@ -1,3 +1,5 @@
+extern alias LethalCompany;
+
 using BeeOverlay.Core.Handlers;
 using BeeOverlay.Core.Ports;
 using BeeOverlay.Core.UseCases;
@@ -5,6 +7,7 @@ using BeeOverlay.Interop;
 using BeeOverlay.Interop.Game;
 using BepInEx.Configuration;
 using BepInEx.Logging;
+using LethalCompany;
 
 namespace BeeOverlay;
 
@@ -15,11 +18,17 @@ internal sealed class PluginController
 {
     private readonly FrameHandler frameHandler;
     private readonly BeeOverlayConfiguration configuration;
+    private readonly HostModPresenceGate hostModPresenceGate;
 
-    private PluginController(FrameHandler frameHandler, BeeOverlayConfiguration configuration)
+    private PluginController(
+        FrameHandler frameHandler,
+        BeeOverlayConfiguration configuration,
+        HostModPresenceGate hostModPresenceGate
+    )
     {
         this.frameHandler = frameHandler;
         this.configuration = configuration;
+        this.hostModPresenceGate = hostModPresenceGate;
     }
 
     public static PluginController Create(ManualLogSource logger, ConfigFile config)
@@ -33,11 +42,39 @@ internal sealed class PluginController
             buildOverlayFrameUseCase: new BuildOverlayFrameUseCase()
         );
 
-        return new PluginController(frameHandler, configuration);
+        return new PluginController(frameHandler, configuration, new HostModPresenceGate());
+    }
+
+    public void AttachHostModPresence(HUDManager hudManager)
+    {
+        hostModPresenceGate.Attach(hudManager);
+    }
+
+    public void ConfirmHostModPresence()
+    {
+        hostModPresenceGate.ConfirmHostPresence();
+    }
+
+    public void BeginHostModPresenceCheck(HostModPresenceBehaviour bridge)
+    {
+        hostModPresenceGate.BeginHostPresenceCheck(bridge);
+    }
+
+    public HostPresenceRequestResult TryRequestHostModPresence()
+    {
+        return hostModPresenceGate.TryRequestHostPresence();
+    }
+
+    public void ResetHostModPresence()
+    {
+        hostModPresenceGate.Reset();
     }
 
     public void HandleFrame()
     {
-        frameHandler.HandleFrame(configuration.Enabled, configuration.PresentationOptions);
+        frameHandler.HandleFrame(
+            configuration.Enabled && hostModPresenceGate.IsOverlayAllowed,
+            configuration.PresentationOptions
+        );
     }
 }